Start tests for summaryFromDocument
