+++ packages/gitbook/package.json
@@ -72,7 +72,8 @@
     "babel-preset-stage-2": "^6.18.0",
     "babel-register": "^6.18.0",
     "expect": "^1.20.2",
-    "mocha": "^3.2.0"
+    "mocha": "^3.2.0",
+    "read-metadata": "^1.0.0"
   },
   "scripts": {
     "test": "./node_modules/.bin/mocha ./testing/setup.js \"./src/**/*/__tests__/*.js\" --bail --reporter=list --timeout=100000 --compilers js:babel-register",
