+++ packages/gitbook/src/parse/summary/__tests__/fixtures/ul-with-link.yaml
@@ -1,0 +1,30 @@
+nodes:
+  - kind: block
+    type: header_one
+    nodes:
+      - kind: text
+        text: Summary
+  - kind: block
+    type: unordered_list
+    nodes:
+      - kind: block
+        type: list_item
+        nodes:
+          - kind: block
+            type: unstyled
+            nodes:
+              - kind: inline
+                type: link
+                data:
+                  href: hello.md
+                nodes:
+                  - kind: text
+                    text: Hello
+      - kind: block
+        type: list_item
+        nodes:
+          - kind: block
+            type: unstyled
+            nodes:
+              - kind: text
+                text: World
